메인현황판 - API 개발 #57 -> 전국전력사용량 타입 double 로 변경 -> 현재 시간 기준으로 금일 이 후 시간은 Double.NaN 으로 처리 하여 차트에 0으로 표기 안되도록 조치

+++ Ami/src/main/java/com/cnu/ami/dashboard/models/UseDayHourAllListVO.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 public class UseDayHourAllListVO {
 
 	private int time;
-	private long use;
+	private double use;
 	private long lp;
 
 }
